pyproject.toml
[tool.poetry]
name = "home_work28"
version = "0.1.0"
description = ""
authors = ["blazkowicz-pnz <ivanbelousov14@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
Pillow = "^9.2.0"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
